requirements.txt
# Core ML Libraries for A6000 GPU
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Transformers and PEFT
transformers>=4.36.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Dataset and Evaluation
datasets>=2.15.0
evaluate>=0.4.0
scikit-learn>=1.3.0

# Visualization and Logging
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.66.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
Pillow>=10.0.0
PyYAML>=6.0.0
psutil>=5.9.0
GPUtil>=1.4.0

# WandB for experiment tracking
wandb>=0.15.0

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Additional dependencies for LLaMA model
sentencepiece>=0.1.99
protobuf>=3.20.0